requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.6.0
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
distro==1.9.0
Django==4.1.5
django-cors-headers==4.3.1
django-crontab==0.7.1
django-environ==0.11.2
django-jazzmin==2.6.0
djangorestframework==3.14.0
frozenlist==1.4.1
geoip2==4.8.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
maxminddb==2.6.2
msgpack==1.0.8
multidict==6.0.5
numpy==2.0.0
openai==1.3.8
packaging==24.1
Pillow==9.5.0
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==24.0.0
pytesseract==0.3.10
pytz==2023.3.post1
redis==5.0.7
requests==2.31.0
service-identity==24.1.0
setuptools==70.1.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tqdm==4.66.1
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2022.7
urllib3==2.1.0
webp==0.4.0
xmltodict==0.13.0
yarl==1.9.4
zope.interface==6.1
